Fix Seperate Card Presentor
